third_party/blink/web_tests/webaudio/decodeAudioData/decode-audio-data-oom-crash.html

<!DOCTYPE html>
<html>
  <head>
    <title>
      Test decodeAudioData Out of Memory Handling
    </title>
    <script src="../../resources/testharness.js"></script>
    <script src="../../resources/testharnessreport.js"></script>
  </head>
  <body>
    <script>
      const createMaliciousWav = () => {
        // Minimal WAV header (44 bytes) with 1Hz sample rate to force massive
        // up-sampling.
        const numSamples = 100000;
        const buffer = new ArrayBuffer(44 + numSamples * 2);
        const view = new DataView(buffer);

        // RIFF
        view.setUint32(0, 0x52494646, false);
        view.setUint32(4, 36 + numSamples * 2, true);
        // WAVE
        view.setUint32(8, 0x57415645, false);
        // fmt
        view.setUint32(12, 0x666d7420, false);
        view.setUint32(16, 16, true); // chunk size
        view.setUint16(20, 1, true); // PCM
        view.setUint16(22, 1, true); // 1 channel
        view.setUint32(24, 1, true); // sample rate: 1 Hz
        view.setUint32(28, 2, true); // byte rate
        view.setUint16(32, 2, true); // block align
        view.setUint16(34, 16, true); // bits per sample
        // data
        view.setUint32(36, 0x64617461, false);
        view.setUint32(40, numSamples * 2, true);

        return buffer;
      };

      promise_test(async (t) => {
        const context = new AudioContext();
        const maliciousBuffer = createMaliciousWav();

        // Attempting to decode a 1Hz file into a standard context demands
        // massive memory (~19GB). This should safely reject via our TryCreate
        // fallback instead of triggering CHECK(bus) crash.
        await promise_rejects_dom(
            t,
            'EncodingError',
            context.decodeAudioData(maliciousBuffer),
            'decodeAudioData should reject with EncodingError when extreme ' +
            'up-sampling memory allocation fails');
      }, 'decodeAudioData safely rejects massive allocations without crashing');
    </script>
  </body>
</html>
